allow initiating websocket connection from any request method
For example, a client may want to POST a multipart file upload request, which then the server can
upgrade the request to a websocket connection to stream processing information back to the client.

This changes the Dial() function to accept a *http.Request so that the user has control over the
request method and the URL. If a user wants to add custom request headers for the Origin or other
headers, those can be attached to the http.Request object through the Header field.